Additional fixes for python 3.10 / PyQt 5.x type handling.
Nils Philippsen posted to the ffado-devel list that additional fixes were
needed beyond what was in r2817.  He pointed out that

  https://github.com/spyder-ide/spyder/issues/16571#issuecomment-940753647

suggests that this might be a PyQt issue rather than python 3.10.  This
patch was supplied by Nils.

As FFADO receives wider testing under python 3.10 and later PyQt 5.x
versions it is possible that further fixes along these lines may be needed.


git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2818 2be59082-3212-0410-8809-b0798e1608f0

+++ support/mixer-qt4/ffado/widgets/matrixmixer.py
@@ -89,9 +89,9 @@ class ColorForNumber:
         lc = self.colors[low]
         hc = self.colors[high]
         return QColor(
-                (1-f)*lc.red()   + f*hc.red(),
-                (1-f)*lc.green() + f*hc.green(),
-                (1-f)*lc.blue()  + f*hc.blue() )
+                int((1-f)*lc.red()   + f*hc.red()),
+                int((1-f)*lc.green() + f*hc.green()),
+                int((1-f)*lc.blue()  + f*hc.blue()) )
 
 class BckgrdColorForNumber(ColorForNumber):
     def __init__(self):
@@ -127,7 +127,7 @@ class MixerNode(QAbstractSlider):
         if max == -1:
             max = pow(2, 16)-1
         self.setRange(0, max)
-        self.setValue(value)
+        self.setValue(int(value))
         self.valueChanged.connect(self.internalValueChanged)
 
         self.setSmall(False)
@@ -693,9 +693,9 @@ class VolumeSlider(QSlider):
         self.setTickPosition(QSlider.TicksBothSides)
         v_min = 10.0*toDBvalue(0)
         v_max = 10.0*toDBvalue(65536)
-        self.setTickInterval((v_max-v_min)/10)
-        self.setMinimum(v_min)
-        self.setMaximum(v_max)
+        self.setTickInterval(int((v_max-v_min)/10))
+        self.setMinimum(int(v_min))
+        self.setMaximum(int(v_max))
         self.setSingleStep(1)
         self.sliderSetValue(value)
         self.In = In
@@ -706,7 +706,7 @@ class VolumeSlider(QSlider):
         #log.debug("Volume slider value changed( %i )" % value)
         v = 10.0*toDBvalue(value)
         #log.debug("Volume slider value changed(dB: %g )" % (0.1*v))
-        self.setValue(v)
+        self.setValue(int(v))
 
     def sliderReadValue(self, value):
         return fromDBvalue(0.1*value)
